# optimize: make inserting host match route several times faster