@membphis 2020-09-10
nice PR

```shell
-- new version
$ time resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua
matched res: 500
route count: 1000
match times: 100000
time used  : 0.62199997901917 sec
QPS        : 160771
resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua  0.64s user 0.00s system 99% cpu 0.646 total

# old version
$ time resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua
matched res: 500
route count: 1000
match times: 100000
time used  : 1.819000005722 sec
QPS        : 54975
resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua  2.00s user 0.01s system 99% cpu 2.020 total
```

2.02 sec -> 0.64 sec.