@spacewander 2020-09-10
@membphis 
Note that the match-hosts.lua in the new version is different from the old one.
To compare the performance, we need to run the new match-hosts.lua with two different versions.