@membphis 2020-09-10
> To compare the performance, we need to run the new match-hosts.lua with two different versions.

```
# master version + match-hosts.lua(this branch)

$ time resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua
matched res: 500
route count: 1000
match times: 100000
time used  : 0.59699988365173 sec
QPS        : 167504
resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua  0.86s user 0.01s system 99% cpu 0.868 total
```