@membphis 2020-09-10
> To compare the performance, we need to run the new match-hosts.lua with two different versions.

ye, you are right. here is the new result.


we use the same `match-hosts.lua` and run it both in the current and `master` branch.

```shell
# this branch
$ time resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua
matched res: 500
route count: 1000
match times: 100000
time used  : 0.62199997901917 sec
QPS        : 160771
resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua  0.64s user 0.00s system 99% cpu 0.646 total

# master branch
$ time resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua
matched res: 500
route count: 1000
match times: 100000
time used  : 0.59699988365173 sec
QPS        : 167504
resty -I=./lib -I=./deps/share/lua/5.1 benchmark/match-hosts.lua  0.86s user 0.01s system 99% cpu 0.868 total
```