@nic-chen 2020-09-10
seems like need to change title like `perf: make inserting host match route several times faster`, and trigger checking again..